edit

+++ ticket_office/main.cpp
@@ -25,7 +25,7 @@ int main()
     }
 
     // определяем скидку в зависимости от возраста самого младшего члена группы
-    float discount = 1 - (min_age / 100);
+    float discount = 1 - ((float)min_age / 100);
 
     // вычисляем общую стоимость билетов
     total_cost = TICKET_PRICE * 5 * discount;
